requirements.txt
[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[packages]
django = "*"
djangorestframework = "*"
gunicorn = "*"
whitenoise = "*"
django-cors-headers = "*"
psycopg2 = "*"

[dev-packages]

[requires]
python_version = "3.9"
